requirements.txt
lxml==4.6.3
Pillow==8.3.1
python-pptx==0.6.19
XlsxWriter==1.4.4
lxml==4.6.3
Pillow==8.3.1
pyodbc==4.0.31
python-pptx==0.6.19
XlsxWriter==1.4.4
lxml==4.6.3
Pillow==8.3.1
pyodbc==4.0.31
python-pptx==0.6.19
six==1.16.0
XlsxWriter==1.4.4
lxml==4.6.3
Pillow==8.3.1
pyodbc==4.0.31
python-pptx==0.6.19
XlsxWriter==1.4.4
